requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
apscheduler==3.10.4
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
